+++ backend/src/rest/UserController.http
@@ -10,7 +10,11 @@ Content-Disposition: form-data; name="name";
 --boundary
 Content-Disposition: form-data; name="nickname";
 
-23
+231
+--boundary
+Content-Disposition: form-data; name="email";
+
+test@test.com
 --boundary
 Content-Disposition: form-data; name="gender";
 
@@ -31,10 +35,11 @@ Content-Disposition: form-data; name="birth";
 Content-Disposition: form-data; name="schoolId";
 
 1
---boundary
-Content-Disposition: form-data; name="avatar"; filename="enjoy.png"
+###--boundary
+#Content-Disposition: form-data; name="avatar"; filename="enjoy.png"
+#
+#< ./../main/resources/mock/images/enjoy.png
 
-< ./../main/resources/mock/images/enjoy.png
 ###
 POST {{apiUrl}}/user/nickname/check
 Authorization: {{authorizationToken}}
@@ -61,6 +66,10 @@ Content-Disposition: form-data; name="name";
 Content-Disposition: form-data; name="nickname";
 
 5
+--boundary
+Content-Disposition: form-data; name="email";
+
+test2@test.com
 --boundary
 Content-Disposition: form-data; name="nicknameChanged";
