+++ backend/src/test/java/com/minibeit/user/domain/UserTest.java
@@ -3,34 +3,28 @@ package com.minibeit.user.domain;
 import com.minibeit.avatar.domain.Avatar;
 import com.minibeit.businessprofile.domain.BusinessProfile;
 import com.minibeit.school.domain.School;
-import com.minibeit.user.dto.AuthRequest;
 import com.minibeit.user.dto.UserRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.mock.web.MockMultipartFile;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("user 도메인 테스트")
 class UserTest {
 
     private User user1;
-    private AuthRequest.Signup signUpRequest;
+    private UserRequest.Signup signUpRequest;
     private School school;
     private Avatar avatar;
     private UserRequest.Update updateRequest;
 
     @BeforeEach
-    void setUp(){
+    void setUp() {
         // given
         avatar = Avatar.builder().id(1L).url("test.url").build();
 
@@ -53,7 +47,7 @@ class UserTest {
     }
 
     private void makeSignupRequest() {
-        signUpRequest = AuthRequest.Signup.builder()
+        signUpRequest = UserRequest.Signup.builder()
                 .name("홍길동")
                 .birth(LocalDate.of(2002, 3, 23))
                 .job("대학생")
@@ -64,7 +58,7 @@ class UserTest {
                 .interestsIds(List.of(1L)).build();
     }
 
-    private void makeUpdateRequest(){
+    private void makeUpdateRequest() {
         updateRequest = UserRequest.Update.builder()
                 .name("바뀐이름")
                 .birth(LocalDate.of(2002, 3, 30))
@@ -80,8 +74,8 @@ class UserTest {
     void signup() {
         user1 = user1.signup(signUpRequest, school, avatar);
         assertAll(
-                ()-> assertThat(user1.getNickname()).isEqualTo(signUpRequest.getNickname()),
-                ()-> assertThat(user1.getName()).isEqualTo(signUpRequest.getName())
+                () -> assertThat(user1.getNickname()).isEqualTo(signUpRequest.getNickname()),
+                () -> assertThat(user1.getName()).isEqualTo(signUpRequest.getName())
         );
     }
 
@@ -91,8 +85,8 @@ class UserTest {
 
         User updatedUser = user1.update(updateRequest, school);
         assertAll(
-                ()-> assertThat(updatedUser.getNickname()).isEqualTo(updateRequest.getNickname()),
-                ()-> assertThat(updatedUser.getName()).isEqualTo(updateRequest.getName())
+                () -> assertThat(updatedUser.getNickname()).isEqualTo(updateRequest.getNickname()),
+                () -> assertThat(updatedUser.getName()).isEqualTo(updateRequest.getName())
         );
     }
 
@@ -109,7 +103,7 @@ class UserTest {
 
     @DisplayName("현재 유저가 비즈니스프로필 어드민인지 확인")
     @Test
-    void isAdminInBusinessProfile(){
+    void isAdminInBusinessProfile() {
         BusinessProfile businessProfile1 = BusinessProfile.builder()
                 .name("동그라미 실험실")
                 .place("고려대")
